Moving PluginVersionsProvider to its correct place, prevents confusion for the plugin implementer that this service is not extending BesuService
