+++ app/src/main/java/org/hyperledger/besu/cli/util/BesuCommandCustomFactory.java
@@ -14,8 +14,6 @@
  */
 package org.hyperledger.besu.cli.util;
 
-import org.hyperledger.besu.plugin.services.PluginVersionsProvider;
-
 import picocli.CommandLine;
 
 /**
