+++ app/src/main/java/org/hyperledger/besu/cli/util/PluginVersionsProvider.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.plugin.services;
+package org.hyperledger.besu.cli.util;
 
 import java.util.Map;
